server: now dish photos are stored in postgre

+++ server/src/main/java/com/ifmo/ratatoile/repository/BusinessLogicRepositories.kt
@@ -28,5 +28,5 @@ interface DishIngredientRepository : JpaRepository<DishIngredient, Int> {
 }
 
 
-interface DishPhotoRepository : JpaRepository<DishPhoto, Int>
+interface DishPhotoRepository : JpaRepository<DishPhoto, Long>
 interface MenuEntryRepository : JpaRepository<MenuEntry, Int>
